Reformating + ESLint
Reformated several loops
Moved "itemCount" const declarations towards the end of the loops

"if (typeof trade.value.rate === 'undefined')" was changed to always assume the trade was a gift.

"if (isGift) trade.prices" was changed to immediately set the object and skip checking if the trade object was missing the "prices" key.

Several other small changes were made so that the code complies with the esLint rules.

The goal here is to make this file more comfortable on the eyes and more accessible for additions in the future.